Reinstate 64-bit Windows builds
So it turns out that a lot of JavaScript code runs up to 2x faster when
ChakraCore is compiled for x64 compared to x86.  While it was nice to
have a sub-7MB installer with only the 32-bit build, I'm not willing to
give up that much performance on 64-bit machines.

+++ src/minisphere/debugger.c
@@ -177,7 +177,7 @@ debugger_update(void)
 			jsal_debug_breakpoint_inject();
 			console_log(0, "connected to debugger at %s", socket_hostname(client));
 			handshake = strnewf("SSj/Ki v%d %s %s\n", KI_VERSION, SPHERE_ENGINE_NAME, SPHERE_VERSION);
-			socket_write(client, handshake, strlen(handshake));
+			socket_write(client, handshake, (int)strlen(handshake));
 			free(handshake);
 			s_socket = client;
 			s_is_attached = true;
